\name{contour2d}
\alias{contour2d}
\title{Display Contour Plot of Mixture}
\description{
Create a contour plot for 2-dimension skewed, cluster weighted, or both mixture }
\usage{
contour2d(opt, method, xlim =NULL, ylim = NULL, mar = c(3.5,3.5,3.5,3.5), nlevels= 100, lwd = 0.2, col = "red", yaxt = NULL, xaxt =NULL)}


\arguments{
  \item{opt}{output object of the function sim.skewcwm().}
  \item{method}{options: "skew", "cwm","skewcwm"}
  \item{xlim}{ x-axis limits}
  \item{ylim}{ y-axis limits }
  \item{mar}{the margin sizes}
  \item{nlevels}{number of contour levels}
  \item{lwd}{line width(s) for the contours drawn.}
  \item{col}{color of contours}
  \item{yaxt}{suppress the y axis}
  \item{xaxt}{suppress the x axis}
}

\details{
...
}







\keyword{contour plot, skewness,cluster weighed model}

